refactor: extract CompoundNode.css.ts from overlays
