+++ packages/diagram/src/overlays/relationships-of/xyflow/CompoundNode.tsx
@@ -5,7 +5,7 @@ import { deepEqual, shallowEqual } from 'fast-equals'
 import { m } from 'framer-motion'
 import { memo, useCallback } from 'react'
 import type { XYFlowTypes } from '../_types'
-import * as css from './styles.css'
+import * as css from '../../shared/xyflow/CompoundNode.css'
 
 const Text = MantineText.withProps({
   component: 'div'
