+++ packages/diagram/src/overlays/relationships-of/xyflow/styles.css.ts
@@ -1,4 +1,3 @@
-import { rem } from '@mantine/core'
 import { globalStyle, style } from '@vanilla-extract/css'
 import { mantine, transitions, vars, whereDark, xyvars } from '../../../theme-vars'
 import { mixColor } from '../../Overlays.css'
@@ -46,28 +45,6 @@ export const elementNodeDescription = style({
   lineHeight: 1.25,
   textWrap: 'balance',
   color: vars.element.loContrast
-})
-
-export const compoundNodeBody = style({
-  position: 'relative',
-  width: '100%',
-  height: '100%',
-  boxShadow: '0 4px 10px 0.5px rgba(0,0,0,0.1) , 1px 1px 4px -1px rgba(0,0,0,0.3)',
-  // background: `color-mix(in srgb , ${vars.element.fill},  transparent 10%)`,
-  background: vars.element.fill,
-  borderRadius: 6
-})
-
-export const compoundNodeTitle = style({
-  fontFamily: vars.compound.font,
-  fontWeight: 600,
-  fontSize: 15,
-  lineHeight: 1,
-  textTransform: 'uppercase',
-  paddingTop: 14,
-  paddingLeft: 12,
-  mixBlendMode: 'screen',
-  color: vars.compound.titleColor
 })
 
 export const cssShapeSvg = style({
